+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `workspaceId` to `formSchema`
+- `workspaceId` to `formSchema` and search options
 
 ## [13.1.0] - 2026-02-23